Refactor - moved the Scm and Issue Management enrichers to standard module Added IT cases for checking the annotations in k8s resources
