+++ enricher/fabric8/src/main/resources/META-INF/fabric8/enricher-default
@@ -9,5 +9,3 @@ io.fabric8.maven.enricher.fabric8.VertxHealthCheckEnricher,507
 io.fabric8.maven.enricher.fabric8.PrometheusEnricher,508
 io.fabric8.maven.enricher.fabric8.AutoTLSEnricher,509
 io.fabric8.maven.enricher.fabric8.ExposeEnricher,1000
-io.fabric8.maven.enricher.fabric8.MavenScmEnricher,510
-io.fabric8.maven.enricher.fabric8.MavenIssueManagementEnricher,511
