+++ it/src/it/simple-maven-issue-mgmt/src/main/java/zero/HelloController.java
@@ -1,0 +1,14 @@
+package zero;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class HelloController {
+
+    @RequestMapping("/")
+    public String index() {
+        return "Greetings from Spring Boot!";
+    }
+
+}
