+++ it/src/it/simple-maven-scm/expected/openshift.yml
@@ -1,0 +1,87 @@
+---
+apiVersion: v1
+kind: List
+items:
+- apiVersion: v1
+  kind: Service
+  metadata:
+    annotations:
+      fabric8.io/git-commit: "@ignore@"
+      prometheus.io/scrape: "true"
+      fabric8.io/git-branch: "@ignore@"
+      prometheus.io/port: "9779"
+      fabric8.io/scm-con-url: "scm:git:git://github.com/fabric8io/fabric8-maven-plugin.git"
+      fabric8.io/scm-devcon-url: "scm:git:git://github.com/fabric8io/fabric8-maven-plugin.git"
+      fabric8.io/scm-tag: "HEAD"
+      fabric8.io/scm-url: "git://github.com/fabric8io/fabric8-maven-plugin.git"
+    labels:
+      expose: "true"
+      provider: fabric8
+      project: fabric8-maven-sample-zero-config
+      version: "@ignore@"
+      group: io.fabric8
+    name: fabric8-maven-sample-zer
+  spec:
+    ports:
+    - name: http
+      port: 80
+      protocol: TCP
+      targetPort: 8080
+    selector:
+      project: fabric8-maven-sample-zero-config
+      provider: fabric8
+      group: io.fabric8
+- apiVersion: v1
+  kind: DeploymentConfig
+  metadata:
+    annotations:
+      fabric8.io/git-commit: "@ignore@"
+      fabric8.io/metrics-path: "@matches('.*dashboard.*fabric8-maven-sample-zero-config.*var-version.*')@"
+      fabric8.io/git-branch: "@ignore@"
+    labels:
+      provider: fabric8
+      project: fabric8-maven-sample-zero-config
+      version: "@ignore@"
+      group: io.fabric8
+    name: fabric8-maven-sample-zer
+  spec:
+    replicas: 1
+    selector:
+      project: fabric8-maven-sample-zero-config
+      provider: fabric8
+      group: io.fabric8
+    template:
+      metadata:
+        annotations:
+          fabric8.io/git-commit: "@ignore@"
+          fabric8.io/metrics-path: "@matches('.*dashboard.*fabric8-maven-sample-zero-config.*var-version.*')@"
+          fabric8.io/git-branch: "@ignore@"
+        labels:
+          provider: fabric8
+          project: fabric8-maven-sample-zero-config
+          version: "@ignore@"
+          group: io.fabric8
+      spec:
+        containers:
+        - env:
+          - name: KUBERNETES_NAMESPACE
+            valueFrom:
+              fieldRef:
+                fieldPath: metadata.namespace
+          image: "@matches('fabric8/fabric8-maven-sample-zero-config:.*$')@"
+          imagePullPolicy: IfNotPresent
+          name: spring-boot
+          ports:
+          - containerPort: 8080
+            name: http
+            protocol: TCP
+          - containerPort: 9779
+            name: prometheus
+            protocol: TCP
+          - containerPort: 8778
+            name: jolokia
+            protocol: TCP
+          securityContext:
+            privileged: false
+    triggers:
+    - type: ConfigChange
